+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Hardening: Sanitize output of callback return value in `_sfce()`
 * Hardening: Sanitize output of `$msg_if_missing` argument value to `_sfcm()`
 * Change: Note compatibility through WP 6.8+
+* Change: Note compatibility through PHP 8.3+
 * Change: Update copyright date (2025)
 
 ## 1.3.3 _(2024-08-09)_